# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API REST para reservas, servicios, combos y administracion de Atrevida Fit.",
        "title": "Atrevida Fit - Agenda API",
        "contact": {},
        "version": "v0.1"
    },
    "basePath": "/",
    "paths": {
...
        "/bd/locales/horarios": {
            "get": {
                "description": "Devuelve los horarios activos de un local con filtro opcional. Query: local_id (requerido), dia_semana 1=lunes a 7=domingo (opcional). Response: total (int), filtros (objeto con local_id, dia_semana), horarios ([]LocalHorarioPG con: id, local_id, dia_semana, hora_desde HH:MM, hora_hasta HH:MM, activo).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Horarios Locales"
                ],
                "summary": "Listar horarios de un local",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 3,
                        "description": "ID del local",
                        "name": "local_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "example": 1,
                        "description": "Dia de la semana (1=lunes, 7=domingo)",
                        "name": "dia_semana",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handlers.horarioListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Error de validacion: local_id invalido, dia_semana debe ser 1-7",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Local no encontrado",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Crea un tramo de horario para un local. Body: local_id (requerido), dia_semana 1-7 (requerido), hora_desde HH:MM (requerido), hora_hasta HH:MM (requerido, debe ser posterior a hora_desde). Response: id (int ID del horario creado).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Horarios Locales"
                ],
                "summary": "Crear horario para un local",
                "parameters": [
                    {
                        "description": "Datos del horario (local_id, dia_semana, hora_desde, hora_hasta)",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.crearLocalHorarioRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handlers.idResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Error de validacion: local_id invalido, dia_semana 1-7, hora_invalida, hora_hasta debe ser posterior",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Local no encontrado",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "409": {
                        "description": "Conflicto: horario ya existe o se superpone con otro existente",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    }
                }
            }
        },
        "/bd/locales/horarios/{id}": {
            "get": {
                "description": "Devuelve un horario por su ID. Param: id (requerido, path). Response: horario (LocalHorarioPG con: id, local_id, dia_semana, hora_desde, hora_hasta, activo).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Horarios Locales"
                ],
                "summary": "Obtener horario por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 15,
                        "description": "ID del horario",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handlers.horarioItemResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Error de validacion: id invalido",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Horario no encontrado",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Realiza borrado logico de un horario (activo=false). Param: id (requerido, path). Response: mensaje string.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Horarios Locales"
                ],
                "summary": "Eliminar horario",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 15,
                        "description": "ID del horario",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handlers.messageResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Error de validacion: id invalido",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Horario no encontrado o ya inactivo",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Actualiza parcialmente un horario. Param: id (requerido, path). Body: dia_semana 1-7 (opcional), hora_desde HH:MM (opcional), hora_hasta HH:MM (opcional, debe ser posterior a hora_desde). Response: mensaje string.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Horarios Locales"
                ],
                "summary": "Actualizar horario",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 15,
                        "description": "ID del horario",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Campos a actualizar (todos opcionales, al menos uno requerido)",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.actualizarLocalHorarioRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handlers.messageResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Error de validacion: id invalido, dia_semana 1-7, hora_invalida, hora_hasta debe ser posterior, sin cambios",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Horario no encontrado",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "409": {
                        "description": "Conflicto: superposicion con otro horario existente",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.actualizarLocalHorarioRequest": {
            "type": "object",
            "properties": {
                "dia_semana": {
                    "type": "integer",
                    "example": 6
                },
                "hora_desde": {
                    "type": "string",
                    "example": "10:00"
                },
                "hora_hasta": {
                    "type": "string",
                    "example": "16:00"
                }
            }
        },
...
        "handlers.crearLocalHorarioRequest": {
            "type": "object",
            "properties": {
                "dia_semana": {
                    "type": "integer",
                    "example": 1
                },
                "hora_desde": {
                    "type": "string",
                    "example": "09:00"
                },
                "hora_hasta": {
                    "type": "string",
                    "example": "18:00"
                },
                "local_id": {
                    "type": "integer",
                    "example": 3
                }
            }
        },
...
        "handlers.horarioFiltrosResponse": {
            "type": "object",
            "properties": {
                "dia_semana": {
                    "type": "integer",
                    "example": 1
                },
                "local_id": {
                    "type": "integer",
                    "example": 3
                }
            }
        },
        "handlers.horarioItemResponse": {
            "type": "object",
            "properties": {
                "horario": {
                    "$ref": "#/definitions/models.LocalHorarioPG"
                }
            }
        },
        "handlers.horarioListResponse": {
            "type": "object",
            "properties": {
                "filtros": {
                    "$ref": "#/definitions/handlers.horarioFiltrosResponse"
                },
                "horarios": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.LocalHorarioPG"
                    }
                },
                "total": {
                    "type": "integer",
                    "example": 5
                }
            }
        },
        "handlers.idResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 42
                }
            }
        },
...
        "handlers.messageResponse": {
            "type": "object",
            "properties": {
                "mensaje": {
                    "type": "string",
                    "example": "operacion realizada correctamente"
                }
            }
        },
...
        "models.LocalHorarioPG": {
            "type": "object",
            "properties": {
                "activo": {
                    "type": "boolean",
                    "example": true
                },
                "dia_semana": {
                    "type": "integer",
                    "example": 1
                },
                "hora_desde": {
                    "type": "string",
                    "example": "09:00"
                },
                "hora_hasta": {
                    "type": "string",
                    "example": "18:00"
                },
                "id": {
                    "type": "integer",
                    "example": 15
                },
                "local_id": {
                    "type": "integer",
                    "example": 3
                }
            }
        },
...
        "utils.APIResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {},
                "error": {
                    "type": "boolean",
                    "example": false
                },
                "message": {
                    "type": "string",
                    "example": "operacion realizada correctamente"
                },
                "status": {
                    "type": "string",
                    "example": "OK"
                }
            }
        }
    }
}